230221/정선아/구매팝업창 확률 수정 + 레시피 목록 수정
1. Evening 월세 차감 수정 - 가지고 있는 돈이 월세보다 적은 경우, 가지고 있는 만큼만 차감
2. EveningScene - 기계구매, 레시피 구매 확률 설정(추후 수정 필요)
3. EveningScene - 목록 갱신 기능 수정, 재화 확인할 수 있게 수정
4. NightScene - 외형 구현
5. DayScene - UserData에 맞춰 초/중/고급 레시피 출력 가능
6. Ingredient, Machine, Recipe 파일 오탈자 수정

+++ Assets/Script/Class/GameData/Machine.cs
@@ -12,7 +12,7 @@ public class Machine
     public int level2Price;
     public int level3Price;
     public string detail;
-    public double level1Pos;
-    public double level2Pos;
-    public double level3Pos;
+    public float level1Pos;
+    public float level2Pos;
+    public float level3Pos;
 }
